requirements.txt
antlr4-python3-runtime==4.7.1
appdirs==1.4.3
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
et-xmlfile==1.0.1
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
jdcal==1.4.1
linecache2==1.0.0
lockfile==0.12.2
msgpack==0.6.2
openpyxl==2.5.2
packaging==20.3
pep517==0.8.2
pkg-resources==0.0.0
progress==1.5
pygame==1.9.6
pyglet==1.5.5
pyparsing==2.4.6
pytoml==0.1.21
requests==2.22.0
retrying==1.3.3
six==1.14.0
traceback2==1.4.0
urllib3==1.25.8
webencodings==0.5.1
